:sparkles: Increase calculator numbers capacity (from float to double) (#16)

+++ src/main/java/com/github/antoinejt/jasc/ConsoleUI.java
@@ -174,7 +174,7 @@ class ConsoleUI {
 
     private static void tryToAddNumberToTheStack(CalculatorEngine calculatorEngine, String input) {
         try {
-            float number = Float.parseFloat(input);
+            double number = Double.parseDouble(input);
             calculatorEngine.addNumber(number);
         } catch (NumberFormatException unused) {
             System.err.println("Your input is invalid!");
